backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.6.3
numpy==1.26.4
